# backend/internal/server/openapi/gateway.openapi.yaml
openapi: 3.1.0
info:
  title: TokenHub Public Gateway API
  summary: OpenAPI contract for TokenHub's externally supported model gateway.
  description: |
    TokenHub exposes OpenAI-compatible, Anthropic-compatible, and native Gemini gateway endpoints.
    Administrative, analytics, health, and metrics endpoints are intentionally excluded from this
    public gateway contract. They require a separate authenticated specification if documented later.

    Entered project API keys are bearer tokens. Documentation clients must keep them in memory,
    never in URLs, local storage, logs, examples, or exported files.
  version: 0.6.0
  x-tokenhub-contract-version: gateway-2026-08-issue-260
...
servers:
  - url: https://tokenhub.example
    description: Replaced at runtime with TOKENHUB_PUBLIC_BASE_URL or the current request origin.
security:
  - TokenHubProjectKey: []
...
paths:
  /v1/models:
    get:
      operationId: listOpenAIModels
      tags: [Models]
      summary: List OpenAI-compatible models.
      description: Returns active models that the project API key is allowed to access.
      x-tokenhub-public-gateway: true
      x-tokenhub-interactive: true
      responses:
        "200":
          description: Accessible models.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAIModelList"
              examples:
                default:
                  value:
                    object: list
                    data:
                      - id: gpt-4.1-mini
                        object: model
                        owned_by: tokenhub
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/models/{model}:
    get:
      operationId: getOpenAIModel
      tags: [Models]
      summary: Get one OpenAI-compatible model.
      description: Returns metadata for one accessible model.
      x-tokenhub-public-gateway: true
      x-tokenhub-interactive: true
      parameters:
        - $ref: "#/components/parameters/ModelPath"
      responses:
        "200":
          description: Model metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAIModel"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/ModelNotFound"
...
  /v1beta/models:
    get:
      operationId: listGeminiModels
      tags: [Gemini Native]
      summary: List native Gemini models.
      description: Returns Gemini-compatible models available to the supplied project API key. Gemini-compatible clients may send the key in `x-goog-api-key`.
      x-tokenhub-public-gateway: true
      x-tokenhub-interactive: true
      security:
        - TokenHubProjectKey: []
        - TokenHubGoogleAPIKey: []
      responses:
        "200":
          description: Gemini model list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiModelList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1beta/models/{model}:
    get:
      operationId: getGeminiModel
      tags: [Gemini Native]
      summary: Get a native Gemini model.
      description: Returns Gemini-compatible model metadata. Gemini-compatible clients may send the TokenHub project API key in `x-goog-api-key`.
      x-tokenhub-public-gateway: true
      x-tokenhub-interactive: true
      security:
        - TokenHubProjectKey: []
        - TokenHubGoogleAPIKey: []
      parameters:
        - $ref: "#/components/parameters/ModelPath"
      responses:
        "200":
          description: Gemini model metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiModel"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/ModelNotFound"
...
components:
  securitySchemes:
    TokenHubProjectKey:
      type: http
      scheme: bearer
      bearerFormat: TokenHub project API key
      description: Project API key generated by TokenHub. Never use an administrator token here.
    TokenHubXAPIKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Protocol-native API key header accepted by Anthropic-compatible clients.
    TokenHubGoogleAPIKey:
      type: apiKey
      in: header
      name: x-goog-api-key
      description: Protocol-native API key header accepted by Gemini-compatible clients and Google Gen AI SDKs.
  parameters:
    ModelPath:
      name: model
      in: path
      required: true
      schema:
        type: string
      examples:
        simple:
          value: gpt-4.1-mini
...
  headers:
    XRequestID:
      description: TokenHub request identifier.
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid project API key.
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ModelNotFound:
      description: Model not found.
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, request_id]
      properties:
        error:
          $ref: "#/components/schemas/ErrorObject"
        request_id:
          type: string
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [invalid_api_key, quota_exceeded, provider_error]
        message:
          type: string
        type:
          type: string
        param:
          type: [string, "null"]
...
    OpenAIModelList:
      type: object
      required: [object, data]
      properties:
        object:
          type: string
          const: list
        data:
          type: array
          items:
            $ref: "#/components/schemas/OpenAIModel"
    OpenAIModel:
      type: object
      required: [id, object]
      properties:
        id:
          type: string
        object:
          type: string
          const: model
        created:
          type: integer
        owned_by:
          type: string
...
    GeminiModelList:
      type: object
      required: [models]
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/GeminiModel"
    GeminiModel:
      type: object
      additionalProperties: true
      properties:
        name:
          type: string
          examples: [models/gemini-2.5-flash]
        displayName:
          type: string
        version:
          type: string
        supportedGenerationMethods:
          type: array
          items:
            type: string
...